Move common code in flow auth to a base class
